Optimize TOO switching between Orders sidebar and Allowances sidebar
