+++ src/pages/Office/MoveOrders/MoveOrders.jsx
@@ -9,7 +9,6 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import styles from './MoveOrders.module.scss';
 
-import DocumentViewer from 'components/DocumentViewer/DocumentViewer';
 import { updateMoveOrder } from 'services/ghcApi';
 import LoadingPlaceholder from 'shared/LoadingPlaceholder';
 import SomethingWentWrong from 'shared/SomethingWentWrong';
@@ -40,7 +39,7 @@ const validationSchema = Yup.object({
 const MoveOrders = () => {
   const history = useHistory();
   const { moveCode } = useParams();
-  const { move, moveOrders, upload, isLoading, isError } = useOrdersDocumentQueries(moveCode);
+  const { move, moveOrders, isLoading, isError } = useOrdersDocumentQueries(moveCode);
   const moveOrderId = move?.ordersId;
 
   const handleClose = () => {
@@ -105,59 +104,50 @@ const MoveOrders = () => {
     sac: moveOrder?.sac,
   };
 
-  const documentsForViewer = Object.values(upload);
-
   return (
-    <div className={styles.MoveOrders}>
-      {documentsForViewer && (
-        <div className={styles.embed}>
-          <DocumentViewer files={documentsForViewer} />
-        </div>
-      )}
-      <div className={styles.sidebar}>
-        <Formik initialValues={initialValues} validationSchema={validationSchema} onSubmit={onSubmit}>
-          {(formik) => (
-            <form onSubmit={formik.handleSubmit}>
-              <div className={styles.orderDetails}>
-                <div className={styles.top}>
-                  <Button
-                    className={styles.closeButton}
-                    data-testid="closeSidebar"
-                    type="button"
-                    onClick={handleClose}
-                    unstyled
-                  >
-                    <FontAwesomeIcon icon="times" title="Close sidebar" aria-label="Close sidebar" />
-                  </Button>
-                  <h2 className={styles.header}>View Orders</h2>
-                  <div>
-                    <Link className={styles.viewAllowances} data-testid="view-allowances" to="allowances">
-                      View Allowances
-                    </Link>
-                  </div>
-                </div>
-                <div className={styles.body}>
-                  <OrdersDetailForm
-                    deptIndicatorOptions={deptIndicatorDropdownOptions}
-                    ordersTypeOptions={ordersTypeDropdownOptions}
-                    ordersTypeDetailOptions={ordersTypeDetailsDropdownOptions}
-                  />
-                </div>
-                <div className={styles.bottom}>
-                  <div className={styles.buttonGroup}>
-                    <Button primary type="submit" disabled={formik.isSubmitting}>
-                      Save
-                    </Button>
-                    <Button type="button" secondary onClick={handleClose}>
-                      Cancel
-                    </Button>
-                  </div>
+    <div className={styles.sidebar}>
+      <Formik initialValues={initialValues} validationSchema={validationSchema} onSubmit={onSubmit}>
+        {(formik) => (
+          <form onSubmit={formik.handleSubmit}>
+            <div className={styles.orderDetails}>
+              <div className={styles.top}>
+                <Button
+                  className={styles.closeButton}
+                  data-testid="closeSidebar"
+                  type="button"
+                  onClick={handleClose}
+                  unstyled
+                >
+                  <FontAwesomeIcon icon="times" title="Close sidebar" aria-label="Close sidebar" />
+                </Button>
+                <h2 className={styles.header}>View Orders</h2>
+                <div>
+                  <Link className={styles.viewAllowances} data-testid="view-allowances" to="allowances">
+                    View Allowances
+                  </Link>
                 </div>
               </div>
-            </form>
-          )}
-        </Formik>
-      </div>
+              <div className={styles.body}>
+                <OrdersDetailForm
+                  deptIndicatorOptions={deptIndicatorDropdownOptions}
+                  ordersTypeOptions={ordersTypeDropdownOptions}
+                  ordersTypeDetailOptions={ordersTypeDetailsDropdownOptions}
+                />
+              </div>
+              <div className={styles.bottom}>
+                <div className={styles.buttonGroup}>
+                  <Button primary type="submit" disabled={formik.isSubmitting}>
+                    Save
+                  </Button>
+                  <Button type="button" secondary onClick={handleClose}>
+                    Cancel
+                  </Button>
+                </div>
+              </div>
+            </div>
+          </form>
+        )}
+      </Formik>
     </div>
   );
 };
